nasiwak_messenger/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
dj-rest-auth==7.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-otp==1.5.4
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
multidict==6.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
twilio==9.4.4
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zope.interface==7.2
